initial is_legal passes perft
bench: 5743954
